refactor: view 配下のファイル整理
