+++ lib/view/widgets/trips/trip_card.dart
